+++ enclave-manager/web/src/components/CodeEditor.tsx
@@ -14,8 +14,15 @@ export const CodeEditor = ({ text, onTextChange, showLineNumbers }: CodeEditorPr
   const isReadOnly = !isDefined(onTextChange);
   const [editor, setEditor] = useState<editor.IStandaloneCodeEditor>();
 
-  const handleContentSizeChange = (e: editor.IContentSizeChangedEvent) => {
-    editor?.layout({ width: 500, height: e.contentHeight });
+  const resizeEditorBasedOnContent = () => {
+    if (isDefined(editor)) {
+      // An initial layout call is needed, else getContentHeight is garbage
+      editor.layout();
+      const contentHeight = editor.getContentHeight();
+      editor.layout({ width: 500, height: contentHeight });
+      // Unclear why layout must be called twice, but seems to be necessary
+      editor.layout();
+    }
   };
 
   const handleMount: OnMount = (editor, monaco) => {
@@ -27,22 +34,27 @@ export const CodeEditor = ({ text, onTextChange, showLineNumbers }: CodeEditorPr
       colors: {},
     });
     monaco.editor.setTheme("kurtosis-theme");
-    editor.onDidContentSizeChange(handleContentSizeChange);
   };
 
   const handleChange: OnChange = (value, ev) => {
     if (isDefined(value) && onTextChange) {
       onTextChange(value);
+      resizeEditorBasedOnContent();
     }
   };
 
+  // Triggering this on every render seems to keep the editor correctly sized
+  // it is unclear why this is the case.
+  resizeEditorBasedOnContent();
+
   return (
-    <Box width={"100%"} minHeight={`${editor?.getContentHeight() || 10}px`}>
+    <Box width={"100%"} maxHeight={"1000px"}>
       <Editor
         onMount={handleMount}
         value={text}
         onChange={handleChange}
         options={{
+          automaticLayout: false, // if this is `true` a ResizeObserver is installed. This causes issues with us managing the container size outside.
           readOnly: isReadOnly,
           lineNumbers: showLineNumbers || (!isDefined(showLineNumbers) && !isReadOnly) ? "on" : "off",
           minimap: { enabled: false },
@@ -59,7 +71,6 @@ export const CodeEditor = ({ text, onTextChange, showLineNumbers }: CodeEditorPr
         }}
         defaultLanguage={"json"}
         theme={"vs-dark"}
-        height={"100%"}
       />
     </Box>
   );