+++ enclave-manager/web/src/components/enclaves/CreateEnclaveButton.tsx
@@ -1,5 +1,5 @@
-import { Button, Menu, MenuButton, MenuItem, MenuList } from "@chakra-ui/react";
-import { FiPackage, FiPlus, FiSettings } from "react-icons/fi";
+import { Button, Menu, MenuButton } from "@chakra-ui/react";
+import { FiPlus } from "react-icons/fi";
 import { useNavigate } from "react-router-dom";
 import { KURTOSIS_CREATE_ENCLAVE_URL_ARG } from "../constants";
 
@@ -8,17 +8,23 @@ export const CreateEnclaveButton = () => {
   return (
     <>
       <Menu matchWidth>
-        <MenuButton as={Button} colorScheme={"kurtosisGreen"} leftIcon={<FiPlus />} size={"md"}>
+        <MenuButton
+          as={Button}
+          colorScheme={"kurtosisGreen"}
+          leftIcon={<FiPlus />}
+          size={"md"}
+          onClick={() => navigate(`#${KURTOSIS_CREATE_ENCLAVE_URL_ARG}`)}
+        >
           New Enclave
         </MenuButton>
-        <MenuList>
-          <MenuItem onClick={() => navigate(`#${KURTOSIS_CREATE_ENCLAVE_URL_ARG}`)} icon={<FiSettings />}>
-            Manual
-          </MenuItem>
-          <MenuItem onClick={() => navigate("/catalog")} icon={<FiPackage />}>
-            Catalog
-          </MenuItem>
-        </MenuList>
+        {/*<MenuList>*/}
+        {/*  <MenuItem onClick={() => navigate(`#${KURTOSIS_CREATE_ENCLAVE_URL_ARG}`)} icon={<FiSettings />}>*/}
+        {/*    Manual*/}
+        {/*  </MenuItem>*/}
+        {/*  <MenuItem onClick={() => navigate("/catalog")} icon={<FiPackage />}>*/}
+        {/*    Catalog*/}
+        {/*  </MenuItem>*/}
+        {/*</MenuList>*/}
       </Menu>
     </>
   );