+++ enclave-manager/web/src/components/enclaves/configuration/inputs/JSONArgumentInput.tsx
@@ -1,25 +1,30 @@
-import { Textarea } from "@chakra-ui/react";
-import { stringifyError } from "../../../../utils";
-import { useEnclaveConfigurationFormContext } from "../EnclaveConfigurationForm";
+import { Controller } from "react-hook-form";
+import { isDefined, stringifyError } from "../../../../utils";
+import { CodeEditor } from "../../../CodeEditor";
 import { KurtosisArgumentTypeInputProps } from "./KurtosisArgumentTypeInput";
 
 export const JSONArgumentInput = (props: Omit<KurtosisArgumentTypeInputProps, "type">) => {
-  const { register } = useEnclaveConfigurationFormContext();
-
   return (
-    <Textarea
-      {...register(props.name, {
-        disabled: props.disabled,
+    <Controller
+      render={({ field }) => <CodeEditor text={field.value} onTextChange={field.onChange} />}
+      name={props.name}
+      defaultValue={"{}"}
+      rules={{
         required: props.isRequired,
-        value: "{}",
         validate: (value: string) => {
           try {
             JSON.parse(value);
           } catch (err: any) {
             return `This is not valid JSON. ${stringifyError(err)}`;
           }
+
+          const propsValidation = props.validate ? props.validate(value) : undefined;
+          if (isDefined(propsValidation)) {
+            return propsValidation;
+          }
         },
-      })}
+      }}
+      disabled={props.disabled}
     />
   );
 };