+++ enclave-manager/web/src/components/enclaves/configuration/utils.ts
@@ -13,7 +13,7 @@ export function argTypeToString(argType?: ArgumentValueType) {
     case ArgumentValueType.LIST:
       return "list";
     case ArgumentValueType.STRING:
-      return "string";
+      return "text";
     default:
       return "json";
   }
@@ -29,7 +29,7 @@ export function argToTypeString(arg: PackageArg) {
     case ArgumentValueType.DICT:
       return `${argTypeToString(arg.typeV2.innerType1)} -> ${argTypeToString(arg.typeV2.innerType2)}`;
     case ArgumentValueType.LIST:
-      return `${argTypeToString(arg.typeV2.innerType1)}[]`;
+      return `${argTypeToString(arg.typeV2.innerType1)} list`;
     default:
       return "json";
   }